02backend/requirements.txt
# Core Application
fastapi==0.111.0
sqlmodel==0.0.19
uvicorn==0.30.1
python-dotenv==1.0.1
pydantic-settings==2.3.4

# Services & APIs
httpx>=0.28.1
tenacity==8.4.1
google-genai==1.23.0
google-generativeai==0.8.3
google-cloud-aiplatform==1.70.0
google-search-results==2.4.2

# CLI & Terminal UI
typer[all]==0.12.3
rich==13.7.1
click==8.1.7

# Database Migration
alembic==1.13.1

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7

# Additional dependencies
pytz==2024.1 